html/browsers/browsing-the-web/navigating-across-documents/015.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<link rel="help" href="https://html.spec.whatwg.org/multipage/links.html#following-hyperlinks">
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#javascript-protocol">
<link rel="help" href="https://html.spec.whatwg.org/multipage/text-level-semantics.html#the-a-element">
<link rel="help" href="https://html.spec.whatwg.org/multipage/webappapis.html#the-event-handler-processing-algorithm">
<link rel="help" href="https://html.spec.whatwg.org/multipage/iframe-embed-object.html#the-iframe-element:about:blank">
<title>Link with onclick and href javascript: url; operations occur in exepected order.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<a target="test" onclick="reports.push('onclick');" href="javascript:parent.onjsurl({data:location.href});">Test</a>
<script>
var test = async_test();
var reports = [];

// This direct sync function call is necessary to avoid
// relying on unrelated events, eg. onload, etc.
window.onjsurl = test.step_func_done(function(e) {
  // This only occurs once.
  reports.push(e.data);

  // "about:blank" is expected last, if this is not received last or
  // not received at all, the test must fail.
  if (e.data === "about:blank") {
    assert_equals(reports[0], "onclick");
    assert_equals(reports[1], "about:blank");
  } else {
    throw new Error("'javascript: url' code is expected to run asynchronously and in the target nested browsing context");
  }
});

test.step(function() {
  document.querySelector("a").click();
});
</script>

<!--
1. This test creates an anchor with an href containing a "javascript: url" .
2. The test clicks the anchor, triggering the "javascript: url" href navigation,
    which must be evaluated and executed asynchronously.
3. `location.href` is used to prove that the "javascript: url"'s contents
    are evaluated in the iframe.
4. If the semantics are implemented correctly, reports will be
    ["onclick", "about:blank"]

-->
